move parseKTX2 code into transcodeKTX2
